requirements.txt
pandas==2.2.2
numpy>=1.22.0
requests>=2.28.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pdf2image>=1.16.0
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.8
faiss-cpu==1.8.0
scikit-learn>=1.0.2
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
crossref-commons>=0.0.7
beautifulsoup4>=4.11.0
lxml>=4.9.0 